requirements.txt
Django==1.10.2
pkg-resources==0.0.0
Django==1.10.2
django-tastypie==0.13.3
pkg-resources==0.0.0
python-dateutil==2.6.0
python-mimeparse==1.6.0
six==1.10.0
